refactor(tui): Move ChannelsView and CostsView to views directory (#1581)
- Move tui/src/components/ChannelsView.tsx to tui/src/views/ChannelsView.tsx
- Move tui/src/components/CostsView.tsx to tui/src/views/CostsView.tsx
- Update imports in app.tsx, test files, and views/index.ts
- Fix relative imports within moved files (../components/AnimatedText, etc.)

This maintains consistent view organization: all view-like components
now reside in the /views directory.
